[IMP] conf: show all relevant versions in the version switcher
From now on, the master branch and the latest released SaaS branch are
always shown. SaaS branches are labelled "Odoo Online". If a newer SaaS
branch is available when the doc is built from another SaaS branch, the
newer branch is labelled "Odoo Online (latest)".

The list of the versions shown in the switcher is hard-coded to force
their ordering.